tests; all fixtures now in conftest.py
moved old tests to new test_models.py

noqa's for some "unused" variables in tests

author/contributor tests and pipeline fixture
